# docs/swagger.yaml
definitions:
...
  api.RoleAdd:
    properties:
      code:
        type: string
      createdAt:
        type: string
      id:
        example: "0"
        type: string
      menuIds:
        items:
          type: string
        type: array
      name:
        type: string
      orgId:
        example: "0"
        type: string
      updatedAt:
        type: string
    type: object
  api.UserAdd:
    properties:
      createdAt:
        type: string
      id:
        example: "0"
        type: string
      orgId:
        example: "0"
        type: string
      password:
        type: string
      realName:
        type: string
      roleIds:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      username:
        type: string
      workNo:
        type: string
    type: object
...
info:
  contact: {}
  title: Admin API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      parameters:
      - description: name 名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: List roles 获取机构列表
      tags:
      - roles 角色
    post:
      consumes:
      - application/json
      parameters:
      - description: Role info 角色信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.RoleAdd'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Create role 创建角色
      tags:
      - roles 角色
  /roles/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Delete role 删除角色
      tags:
      - roles 角色
    get:
      consumes:
      - application/json
      parameters:
      - description: Org ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Role 获取角色
      tags:
      - roles 角色
    put:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Role info 角色信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.RoleAdd'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Update role 更新角色
      tags:
      - roles 角色
  /roles/{id}/menus:
    get:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Role menus 获取角色菜单
      tags:
      - roles 角色
  /roles/orgs/{orgId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Org ID
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Org roles 获取机构角色
      tags:
      - roles 角色
  /users:
    get:
      consumes:
      - application/json
      parameters:
      - description: name 名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: List users 用户列表
      tags:
      - users 用户
    post:
      consumes:
      - application/json
      parameters:
      - description: User info 用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.UserAdd'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Create user 创建用户
      tags:
      - users 用户
  /users/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Delete user 删除用户
      tags:
      - users 用户
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: User 获取用户
      tags:
      - users 用户
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User info 用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.UserAdd'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Update users 更新用户
      tags:
      - users 用户
  /users/{id}/roles:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: User roles 获取用户角色
      tags:
      - users 用户
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"